Moved aws access functions and render_template from __init__.py into it's own files
