+++ exasol_script_languages_container_ci_setup/lib/source_credentials.py
@@ -1,7 +1,8 @@
 import logging
 
 import exasol_script_languages_container_ci_setup
-from exasol_script_languages_container_ci_setup.lib import render_template, validate_cloudformation_template
+from exasol_script_languages_container_ci_setup.lib.aws_access import validate_cloudformation_template
+from exasol_script_languages_container_ci_setup.lib.render_template import render_template
 
 SOURCE_CREDENTIALS_STACK_NAME = "SLCSourceCredentials"
 
@@ -19,7 +20,7 @@ def run_deploy_source_credentials(
                           secret_name=secret_name,
                           secret_user_key=secret_user_key,
                           secret_token_key=secret_token_key)
-    exasol_script_languages_container_ci_setup.lib.upload_cloudformation_stack(aws_profile, yml,
+    exasol_script_languages_container_ci_setup.lib.aws_access.upload_cloudformation_stack(aws_profile, yml,
                                                                                SOURCE_CREDENTIALS_STACK_NAME)
