+++ test/test_generate_buildspec.py
@@ -3,7 +3,7 @@ import json
 import jsonschema
 import pytest
 
-from exasol_script_languages_container_ci_setup.lib import render_template
+from exasol_script_languages_container_ci_setup.lib.render_template import render_template
 from exasol_script_languages_container_ci_setup.lib.run_generate_buildspec import run_generate_buildspec, \
     get_config_file_parameter
